Add files via upload
Added nursery
Changed vore; fullness now counts current weight again so I could use "var content" to check if current content is more than maxcapacity, and if so regurgitate the excess prey doing so I hopefully got rid of the divided by zero bug is capacity is zero. 
changed the switch statement for item in checkflags to a for loop.